Create a requirements.txt and a pythonapp.yaml file so that the github actions will work properly
Requirements.txt was created by making a fresh python environment and trying to run main.py (while importing all of the classes) and then pip installing all of the dependencies as they came up. The yaml file was created using the template from Logan's Software QA class, this file was originally meant to lint with flake8 but needs to be refactored to lint with pylint (if that remains to be the intention). Requirements.txt needs to be kept up-to-date as the team adds more requirements to the project.