Moved predict_mask functions into happy_transformer and cleaned them up.
Moved the functions and fixed up formatting issues. Merged predict_mask and predict_mask_with_options in order to completely remove duplication of code between the two functions. Also tried to merge helper functions for the predict_mask into resonable sub methods and group them together. Added a note for future that developers should be overloading this method if they are used in future transformers that do not need this style to predict a mask.